Hoist byte-identical truth-table cell pairs into shared bases (#27 S4)
The S4 diff-audit found InputSig/OutputSig and InputVal/OutputVal to be
clones of one another (OutputVal even ended with '// end of InputVal
class'). Both pairs now sit on small abstract bases:

- SigEntry owns the signal-name field, sizing, drawing, and the
  delete/rename/move popup; the four menu actions dispatch through
  template hooks (doRemove/doRename/doMoveLeft/doMoveRight), so each
  subclass is just the four one-line TruthTable calls it always
  differed by. No input/output flag anywhere.
- ValEntry owns the value field, sizing, and 0/1/x drawing; selected()
  stays in the subclasses because their click behaviors genuinely
  differ (don't-care expand/collapse vs output toggle).

The remaining audited pairs (InputPin/OutputPin residuals into the
existing Pin base; JumpStart/JumpEnd into a new Jump base) are
documented on the issue with per-method hoistability verdicts.
