+++ src/jls/elem/OutputVal.java
@@ -1,18 +1,14 @@
 package jls.elem;
 
-import java.awt.Color;
-import java.awt.FontMetrics;
 import java.awt.Graphics;
 
 /**
- * Draw an output signal value and react to value being clicked on in the truth table.
+ * An output signal value in a truth table; drawing and sizing live in
+ * ValEntry (#27 S4).
  * 
  * @author David A. Poplawski
  */
-public final class OutputVal extends Entry {
-	
-	// properties
-	int value = 0;  // 0 or 1, 2 = don't care
+public final class OutputVal extends ValEntry {
 
 	/**
 	 * Create a new entry.
@@ -23,30 +19,8 @@ public final class OutputVal extends Entry {
 	 */
 	public OutputVal(TruthTable ttelem, int value, Graphics g) {
 		
-		super(ttelem);
-		this.value = value;
-		FontMetrics fm = g.getFontMetrics();
-		minHeight = fm.getAscent()+fm.getDescent();
-		minWidth = fm.stringWidth(" " + value + " ");
+		super(ttelem,value,g);
 	} // end of constructor
-
-	/**
-	 * Draw this entry.
-	 * 
-	 * @param g The Graphics object to draw with.
-	 */
-	public void draw(Graphics g) {
-
-		FontMetrics fm = g.getFontMetrics();
-		int ascent = fm.getAscent();
-		int height = ascent + fm.getDescent();
-		String str = "" + value;
-		if (value == 2)
-			str = "x";
-		int width = fm.stringWidth(str);
-		g.setColor(Color.BLACK);
-		g.drawString(str,x+(this.width-width)/2,y+(this.height-height)/2+ascent);
-	} // end of draw method
 
 	/**
 	 * Change value when selected.
@@ -59,5 +33,4 @@ public final class OutputVal extends Entry {
 		ttelem.toggleOutput(row-2,col-1);
 	} // end of selected method
 
-} // end of InputVal class
-
+} // end of OutputVal class
